% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_combine.R
\name{pc_combine}
\alias{pc_combine}
\title{Combine whitelisted sources and referrers with additional data extracted from a website}
\usage{
pc_combine(...)
}
\arguments{
\item{...}{Named functions}
}
\value{
Nothing, used for its side effects (fills relevant column in the `combine` spreadsheet)
}
\description{
Combine whitelisted sources and referrers with additional data extracted from a website
}
\examples{
\dontrun{
pc_combine(url = "https://www.europeandatajournalism.eu/eng/News/Data-news/The-price-of-coastal-flood-mitigation-in-Europe",
           title = function(url) {
             pc_extract_from_web(url = url,
                                 container = "h1")
           },
           date = function(url) {
             pc_extract_from_web(url = url,
                                 container = "div",
                                 container_class = "date") \%>\%
               stringr::str_remove(pattern = "[[:alpha:]]+") \%>\%
               anytime::anydate()}
)
}

}
