% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_downlodad_backlinks.R
\name{pc_downlodad_backlinks}
\alias{pc_downlodad_backlinks}
\title{Downloads backlinks stored in Google Drive}
\usage{
pc_downlodad_backlinks(date_range = c(Sys.Date() - 32, Sys.Date() - 1))
}
\arguments{
\item{date_range}{A vector of length two, defaults to c(Sys.Date()-32, Sys.Date()-1).}
}
\value{

}
\description{
Downloads backlinks stored in Google Drive
}
\examples{
\dontrun{
pc_downlodad_backlinks()
}
}
