% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_find_dribble.R
\name{pc_find_dribble}
\alias{pc_find_dribble}
\title{Find dribbles used by popularitycheckr}
\usage{
pc_find_dribble(
  year = NULL,
  type = "raw_data",
  content = "ga_backlinks",
  base_folder = "popularitycheckr",
  create_if_missing = TRUE
)
}
\arguments{
\item{year}{A vector of length one. Defaults to NULL, to be used only if relevant.}

\item{type}{Type of data that the dribble refers to. Defaults to "raw_data".}

\item{content}{Describes what contents will be included in the given spreadsheet. Defaults to "ga_backlinks".}

\item{base_folder}{Defaults to "popularitycheckr". All files and folders will be included inside this folder.}

\item{create_if_missing}{Defaults to TRUE. If already existing, it just returns the relevant dribble.}
}
\value{
A dribble
}
\description{
Find dribbles used by popularitycheckr
}
\examples{

\dontrun{
pc_find_dribble(2020)
}
}
