% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_get_backlinks.R
\name{pc_get_backlinks}
\alias{pc_get_backlinks}
\title{Get referral backlinks to my website.}
\usage{
pc_get_backlinks(
  date_range = c(Sys.Date() - 32, Sys.Date() - 1),
  sources_to_exclude = popularitycheckr::sources_to_exclude,
  ga_id = NULL
)
}
\arguments{
\item{date_range}{A characther vector of length two, providing the first and the last day to include in the data request, e.g. `c("2020-06-01", "2020-06-30")`.}

\item{sources_to_exclude}{A characther vector of domains to be excluded. By default, a vector of commonly found search engines, social media, and webmail providers. See popularitycheckr::sources_to_exclude}

\item{ga_id}{A Google Analytics viewId. Defaults to NULL. If not given, it uses the environment variable typically set with `pc_set_ga_id()`.}
}
\value{
A data frame (a tibble) of four columns:
}
\description{
Get referral backlinks to my website.
}
\examples{

\dontrun{
pc_set_ga_id(1234567)

backlinks <- pc_get_backlinks(date_range = c(Sys.Date()-32,
Sys.Date()-1))
backlinks

}
}
