% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_get_ia.R
\name{pc_get_ia_url}
\alias{pc_get_ia_url}
\title{Gets an Archive.org Wayback Machine URL}
\usage{
pc_get_ia_url(url)
}
\arguments{
\item{url}{A charachter vector of length one, a url.}
}
\value{
A url linking to the version on the Internet Archive
}
\description{
See https://archive.org/help/wayback_api.php; and https://github.com/hrbrmstr/wayback for an R implementation.
}
\examples{
\dontrun{

}
}
