% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_process_backlinks}
\alias{pc_process_backlinks}
\title{Process backlinks}
\usage{
pc_process_backlinks(backlinks, include_domain = TRUE)
}
\arguments{
\item{backlinks}{A data frame with backlinks, typically generated with `pc_get_backlinks()` or retrieved with `pc_downlodad_backlinks()`}

\item{include_domain}{Logical, defaults to TRUE. Include the base domain in the `landingPagePath` column for clarity, assuming https protocol.}
}
\value{

}
\description{
Uploads backliks to relevant google drive spreadsheet, consdering custom blacklist and whitelists.
}
\examples{
\dontrun{
pc_process_backlinks(pc_get_backlinks())
}
}
