% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_process_inputs}
\alias{pc_process_inputs}
\title{Process manual inputs}
\usage{
pc_process_inputs(columns = c("public", "dismiss"), spreadsheet = "combine")
}
\arguments{
\item{lists}{A characther vector giving the names of the categories for manual categorisation. Defaults to `c("blacklist_source","blacklist_referrer","whitelist_source","whitelist_referrer")`.}
}
\value{
Nothing, used for its side effects.
}
\description{
Takes manual inputs, records them in lists in other sheets of the same worksheet, and removes them from the "process" sheet.
}
\examples{
\dontrun{
pc_process_inputs()
}
}
