% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_process_lists}
\alias{pc_process_lists}
\title{Process whitelists/blacklists or other custom lists}
\usage{
pc_process_lists(
  lists = c("blacklist_source", "blacklist_referrer", "whitelist_source",
    "whitelist_referrer")
)
}
\arguments{
\item{lists}{A characther vector giving the names of the categories for manual categorisation. Defaults to `c("blacklist_source","blacklist_referrer","whitelist_source","whitelist_referrer")`.}
}
\value{
Nothing, used for its side effects.
}
\description{
Takes manual inputs, records them in lists in other sheets of the same worksheet, and removes them from the "process" sheet.
}
\examples{
\dontrun{
pc_process_lists()
}
}
