% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_process_move_whitelisted}
\alias{pc_process_move_whitelisted}
\title{Move whitelisted backlinks to a separate spreadsheet to be combined with other data}
\usage{
pc_process_move_whitelisted(backlinks)
}
\arguments{
\item{backlinks}{A data frame with backlinks, typically generated with `pc_get_backlinks()` or retrieved with `pc_downlodad_backlinks()`}
}
\value{

}
\description{
Move whitelisted backlinks to a separate spreadsheet to be combined with other data
}
\examples{
\dontrun{
pc_process_move_whitelisted(pc_get_backlinks())
googledrive::drive_browse(pc_find_dribble(type = "inputs", content = "combine"))
}
}
