% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_set_domain.R
\name{pc_set_domain}
\alias{pc_set_domain}
\title{Set or retrieve the domain you are currently working with as an environment variable}
\usage{
pc_set_domain(domain = NULL)
}
\arguments{
\item{domain}{A plain domain name, such as europeandatajournalism.eu}
}
\value{
The domain currently recorded as an environment variable.
}
\description{
Set or retrieve the domain you are currently working with as an environment variable
}
\examples{

pc_set_domain(domain = "europeandatajournalism.eu")

# If set, this function can be used to retrieve the id.

my_domain <- pc_set_domain()
my_domain
}
