% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_set_ga_id.R
\name{pc_set_ga_id}
\alias{pc_set_ga_id}
\title{Set or retrieve the Google Analytics viewId as an environment variable}
\usage{
pc_set_ga_id(ga_id = NULL)
}
\arguments{
\item{ga_id}{A numeric vector of length one corresponding to a Google Analyitcs viewId.}
}
\value{
The Google Analytics viewId currently recorded as environment variable.
}
\description{
Set or retrieve the Google Analytics viewId as an environment variable
}
\examples{

pc_set_ga_id(ga_id = 1234567)

# If set, this function can be used to retrieve the id.

my_ga_id <- pc_set_ga_id()
my_ga_id
}
