% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_upload_backlinks.R
\name{pc_upload_backlinks}
\alias{pc_upload_backlinks}
\title{Uploads ga backlinks to Google Drive}
\usage{
pc_upload_backlinks(backlinks)
}
\arguments{
\item{backlinks}{}
}
\value{

}
\description{
Uploads ga backlinks to Google Drive
}
\examples{
\dontrun{
pc_set_ga_id(1234567)

pc_get_backlinks() \%>\%
  pc_upload_backlinks()

}
}
