% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_clean_lists}
\alias{pc_clean_lists}
\title{Removes elements already whitelisted or blacklisted}
\usage{
pc_clean_lists(
  lists = c("blacklist_source", "blacklist_referrer", "whitelist_source",
    "whitelist_referrer"),
  spreadsheet = "process"
)
}
\arguments{
\item{spreadsheet}{}
}
\value{

}
\description{
Removes elements already whitelisted or blacklisted
}
\examples{
\dontrun{
pc_clean_lists()
}
}
