% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_process_backlinks.R
\name{pc_process_backlinks}
\alias{pc_process_backlinks}
\title{Process backlinks}
\usage{
pc_process_backlinks(
  backlinks = NULL,
  lists = c("blacklist_source", "blacklist_referrer", "whitelist_source",
    "whitelist_referrer")
)
}
\arguments{
\item{backlinks}{A data frame with backlinks, typically generated with `pc_get_backlinks()` or retrieved with `pc_downlodad_backlinks()`}

\item{lists}{A characther vector giving the names of the categories for manual categorisation. Defaults to `c("blacklist_source","blacklist_referrer","whitelist_source","whitelist_referrer")`.}
}
\value{

}
\description{
Uploads backliks to relevant google drive spreadsheet, consdering custom blacklist and whitelists.
}
\examples{
\dontrun{
pc_prepare_backlinks(pc_get_backlinks())
}
}
